Rework the site into a static site generated with Hugo.
The database is super simple, so can work really easily as just a static
JSON file.  This will be faster to serve static files, instead of the
current, pretty slow, appengine site.

Also, will be easier to use GitHub's collaboration features to track and
vet changes to the database and add co-maintainers, without having to
implement that with code.